kisan-sahayak/requirements.txt
streamlit
streamlit-option_menu
google-generativeai
edge-tts
Pillow
requests
streamlit-mic-recorder